+++ api-augment/substrate/registry.ts
@@ -1,2 +1,2 @@
 
-import 'https://deno.land/x/polkadot@0.2.39/types-augment/registry/substrate.ts';
+import 'https://deno.land/x/polkadot/types-augment/registry/substrate.ts';