+++ api-derive/session/info.ts
@@ -4,7 +4,7 @@ import type { DeriveApi, DeriveSessionInfo } from '../types.ts';
 
 import { map } from 'https://esm.sh/rxjs@7.8.1';
 
-import { objectSpread } from 'https://deno.land/x/polkadot@0.2.39/util/mod.ts';
+import { objectSpread } from 'https://deno.land/x/polkadot/util/mod.ts';
 
 import { memo } from '../util/index.ts';