+++ api/detectOther.ts
@@ -1,8 +1,8 @@
 
-import { packageInfo as deriveInfo } from 'https://deno.land/x/polkadot@0.2.39/api-derive/packageInfo.ts';
-import { packageInfo as coreInfo } from 'https://deno.land/x/polkadot@0.2.39/rpc-core/packageInfo.ts';
-import { packageInfo as providerInfo } from 'https://deno.land/x/polkadot@0.2.39/rpc-provider/packageInfo.ts';
-import { packageInfo as typesInfo } from 'https://deno.land/x/polkadot@0.2.39/types/packageInfo.ts';
-import { packageInfo as knownInfo } from 'https://deno.land/x/polkadot@0.2.39/types-known/packageInfo.ts';
+import { packageInfo as deriveInfo } from 'https://deno.land/x/polkadot/api-derive/packageInfo.ts';
+import { packageInfo as coreInfo } from 'https://deno.land/x/polkadot/rpc-core/packageInfo.ts';
+import { packageInfo as providerInfo } from 'https://deno.land/x/polkadot/rpc-provider/packageInfo.ts';
+import { packageInfo as typesInfo } from 'https://deno.land/x/polkadot/types/packageInfo.ts';
+import { packageInfo as knownInfo } from 'https://deno.land/x/polkadot/types-known/packageInfo.ts';
 
 export default [deriveInfo, coreInfo, providerInfo, typesInfo, knownInfo];