+++ rpc-core/types/index.ts
@@ -1,4 +1,4 @@
 
-export * from 'https://deno.land/x/polkadot@0.2.39/rpc-core/types/jsonrpc.ts';
+export * from 'https://deno.land/x/polkadot/rpc-core/types/jsonrpc.ts';
 
 export * from './base.ts';