+++ rpc-provider/detectOther.ts
@@ -1,4 +1,4 @@
 
-import { packageInfo as typesInfo } from 'https://deno.land/x/polkadot@0.2.39/types/packageInfo.ts';
+import { packageInfo as typesInfo } from 'https://deno.land/x/polkadot/types/packageInfo.ts';
 
 export default [typesInfo];