+++ types-augment/lookup/definitions.ts
@@ -1,7 +1,7 @@
 
-import type { Definitions } from 'https://deno.land/x/polkadot@0.2.39/types/types/index.ts';
+import type { Definitions } from 'https://deno.land/x/polkadot/types/types/index.ts';
 
-import { objectSpread } from 'https://deno.land/x/polkadot@0.2.39/util/mod.ts';
+import { objectSpread } from 'https://deno.land/x/polkadot/util/mod.ts';
 
 import kusama from './kusama.ts';
 import polkadot from './polkadot.ts';