+++ types-codec/base/Null.ts
@@ -1,8 +1,8 @@
 
-import type { HexString } from 'https://deno.land/x/polkadot@0.2.39/util/types.ts';
+import type { HexString } from 'https://deno.land/x/polkadot/util/types.ts';
 import type { Codec, Inspect, IU8a, Registry } from '../types/index.ts';
 
-import { isNull } from 'https://deno.land/x/polkadot@0.2.39/util/mod.ts';
+import { isNull } from 'https://deno.land/x/polkadot/util/mod.ts';
 
 /**
  * @name Null