+++ types-codec/test/performance.ts
@@ -1,6 +1,8 @@
 
+/* global it */
 
-import { formatDecimal, formatNumber } from 'https://deno.land/x/polkadot@0.2.39/util/mod.ts';
+
+import { formatDecimal, formatNumber } from 'https://deno.land/x/polkadot/util/mod.ts';
 
 type ExecFn = (...params: any[]) => unknown;