+++ types-create/bundle.ts
@@ -2,5 +2,4 @@
 export { packageInfo } from './packageInfo.ts';
 export { TypeDefInfo } from './types/index.ts';
 
-export * from './create/index.ts';
-export * from './util/index.ts';
+export * from './exports.ts';