+++ types-create/packageInfo.ts
@@ -1,3 +1,3 @@
 
 
-export const packageInfo = { name: '@polkadot/types-create', path: new URL(import.meta.url).pathname, type: 'deno', version: '0.2.39' };
+export const packageInfo = { name: '@polkadot/types-create', path: new URL(import.meta.url).pathname, type: 'deno', version: '10.7.2' };