+++ types/types/codec.ts
@@ -1,7 +1,7 @@
 
-import type { Codec } from 'https://deno.land/x/polkadot@0.2.39/types-codec/types/index.ts';
+import type { Codec } from 'https://deno.land/x/polkadot/types-codec/types/index.ts';
 
-export type { AnyFunction, AnyJson, AnyNumber, AnyString, AnyTuple, AnyU8a, ArgsDef, BareOpts, Codec, CodecClass, CodecTo, CodecClass as Constructor, Inspect } from 'https://deno.land/x/polkadot@0.2.39/types-codec/types/index.ts';
+export type { AnyFunction, AnyJson, AnyNumber, AnyString, AnyTuple, AnyU8a, ArgsDef, BareOpts, Codec, CodecClass, CodecTo, CodecClass as Constructor, Inspect } from 'https://deno.land/x/polkadot/types-codec/types/index.ts';
 
 export type ArrayElementType<T extends ReadonlyArray<unknown>> = T extends ReadonlyArray<infer ElementType>
   ? ElementType